requirements.txt
-i https://pypi.org/simple

anyio==3.6.1

apscheduler==3.9.1

cachecontrol==0.12.11

cachetools==5.2.0

certifi==2022.6.15

charset-normalizer==2.0.12

firebase-admin==5.2.0

google-api-core[grpc]==2.8.2

google-api-python-client==2.51.0

google-auth==2.8.0

google-auth-httplib2==0.1.0

google-cloud-core==2.3.1

google-cloud-firestore==2.5.3

google-cloud-storage==2.4.0

google-crc32c==1.3.0

google-resumable-media==2.3.3

googleapis-common-protos==1.56.3

grpcio==1.47.0

grpcio-status==1.47.0

h11==0.12.0

httpcore==0.15.0

httplib2==0.20.4

httpx==0.23.0

idna==3.3

msgpack==1.0.4

proto-plus==1.20.6

protobuf==3.20.1

pyasn1==0.4.8

pyasn1-modules==0.2.8

pyparsing==3.0.9

python-telegram-bot==20.0a1

pytz==2022.1

pytz-deprecation-shim==0.1.0.post0

requests==2.28.0

rfc3986[idna2008]==1.5.0

rsa==4.8

setuptools==62.6.0

six==1.16.0

sniffio==1.2.0

tornado==6.2b2

tzdata==2022.1

tzlocal==4.2

uritemplate==4.1.1

urllib3==1.26.9

